refactor: 코드 재활용을 위한 패키지 이동
